requirement.txt
fitz
pymupdf
pandas
regex
langchain==1.0.2
langchain-core==1.0.0
langchain-google-genai==3.0.0
python-dotenv==1.1.1
pydantic==2.12.3